# Api_Gateway/api/handler/budgeting/notification.go
package handler

import (
	pb "api/genprotos/budgeting"
	"context"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
)

// @Summary ListNotifications
// @Description ListNotifications
// @Tags Notification
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param limit query int false "Limit"
// @Param page query int false "Page"
// @Success 200 {object} pb.GetNotificationResponse
// @Failure 400 {string} string "Bad Request"
// @Router /notification/get [get]
func (h *BudgetingHandler) ListNotifications(ctx *gin.Context) {

	defaultLimit := 10
	defaultPage := 1

	limitStr := ctx.Query("limit")
	pageStr := ctx.Query("page")

	limit, err := strconv.Atoi(limitStr)
	if err != nil || limit <= 0 {
		limit = defaultLimit
	}

	page, err := strconv.Atoi(pageStr)
	if err != nil || page <= 0 {
		page = defaultPage
	}

	req := &pb.GetNotificationRequest{
		Limit: int32(limit),
		Page:  int32(page),
	}

	res, err := h.Notification.GetNotification(context.Background(), req)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, err.Error())
		return
	}

	ctx.JSON(http.StatusOK, res)
}